removing warnings
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4294 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/graphics/Model3D.hpp
@@ -156,7 +156,7 @@ namespace graphics {
 
             void addTriangles(const std::vector<rw::geometry::IndexedTriangle<uint16_t> >& tris){
                 //TODO Is it safe to assume that we don't have more than 65536 faces? 
-				uint16_t startIdx = _faces.size();
+				uint16_t startIdx = (uint16_t) _faces.size();
                 _faces.resize(_faces.size()+tris.size());
                 for(size_t i=0;i<tris.size();i++){
                     _faces[startIdx+i] = tris[i];
